Trend filter controls and additional KPIs
1. Implemented remaining filter controls to the trend visual. Verified
individual filters and combination of filters.  They work as expected.
2. Added % O/S, Default Propensity, and Delinquency Recency to the KPI
tab.  The numbers match Zoomdata when selecting a Loan Grade filter.
However, when unfiltered, it needs to obtain KPI values with no
grouping.  Adding up the KPI for all grades is not correct.
Investigating if totals can be obtained from the same API call or if
this needs an additional WS request.